Split starlite/cli.py into starlite/cli + submodules
